Add man page
I find myself always forgetting what each column means, and felt compelled to
write a man page for elasticstat(1). They're generated using `ronn`, because
writing a man page by hand is a form of masochism which I choose not to endure.

Sadly, installation of the man page is safest when done by hand, but it's still
good to have hanging around in case we ever decide to package it up in an rpm
or deb at some point.